First challenge...
The problem is quite a common one from the offset and trying to split down brackets. The difference is that there is a number prefix

Should start with a non-optimal solution, count for the opened brackets. When they are equal then I can snip that section out. Inside that element I can break it down further